pyproject.toml
[tool.poetry]
name = "repl_nix_HalalBoy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^2.1.0"
python-dotenv = "^0.21.0"
discord-py = "^2.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
